feat(config): add config v7 with optional destinations and per-destination targets
Makes [destinations] optional (v7 no longer requires at least one, unlike
v3-v6) and lets a destination pin its own gnosis_vpn_server/wireguard_server,
overriding the global [connection.bridge/wg].target default when present -
laying the groundwork for exit-node discovery to populate destinations
without any static config at all.

v7 becomes the canonical schema; v3-v6 now hold only their own historical
dialect differences and forward-convert into v7::Config instead of each
duplicating the shared connection/wireguard/blokli/strategy schema and
converting straight to the runtime config. Destination gains a
DestinationSource tag (Configured/Discovered/ConfiguredAndDiscovered) so a
later discovery merge can tell the two origins apart.
