+++ gnosis_vpn-lib/src/command/balance_response.rs
@@ -119,7 +119,7 @@ impl Display for ChannelBalance {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::connection::destination::{Destination, HopRouting};
+    use crate::connection::destination::{Destination, DestinationSource, HopRouting};
 
     fn address(byte: u8) -> Address {
         Address::from([byte; 20])
@@ -131,6 +131,9 @@ mod tests {
             addr,
             HopRouting::try_from(1).expect("conversion cannot fail"),
             HashMap::new(),
+            None,
+            None,
+            DestinationSource::Configured,
         )
     }
