+++ gnosis_vpn-lib/src/config/mod.rs
@@ -15,6 +15,7 @@ mod v3;
 mod v4;
 mod v5;
 mod v6;
+mod v7;
 
 pub const DEFAULT_PATH: &str = "/etc/gnosisvpn/config.toml";
 pub const ENV_VAR: &str = "GNOSISVPN_CONFIG_PATH";
@@ -40,8 +41,6 @@ pub enum Error {
     TomlDeserialization(#[from] toml::de::Error),
     #[error("Unsupported config version: {0}")]
     VersionMismatch(u8),
-    #[error("No destinations")]
-    NoDestinations,
     #[error("ping and main sessions must both have surb_balancing enabled or both disabled")]
     SurbBalancingMismatch,
     #[error("Error in hopr-lib: {0}")]
@@ -70,6 +69,7 @@ pub async fn read(path: &Path) -> Result<Config, Error> {
             for key in wrong_keys.iter() {
                 tracing::warn!(%key, "ignoring unsupported key in configuration file");
             }
+            let res: v7::Config = res.try_into()?;
             res.try_into()
         }
         4 => {
@@ -78,6 +78,7 @@ pub async fn read(path: &Path) -> Result<Config, Error> {
             for key in wrong_keys.iter() {
                 tracing::warn!(%key, "ignoring unsupported key in configuration file");
             }
+            let res: v7::Config = res.try_into()?;
             res.try_into()
         }
         5 => {
@@ -86,11 +87,21 @@ pub async fn read(path: &Path) -> Result<Config, Error> {
             for key in wrong_keys.iter() {
                 tracing::warn!(%key, "ignoring unsupported key in configuration file");
             }
+            let res: v7::Config = res.try_into()?;
             res.try_into()
         }
         6 => {
             let res = toml::from_str::<v6::Config>(&content)?;
             let wrong_keys = v6::wrong_keys(&table);
+            for key in wrong_keys.iter() {
+                tracing::warn!(%key, "ignoring unsupported key in configuration file");
+            }
+            let res: v7::Config = res.try_into()?;
+            res.try_into()
+        }
+        7 => {
+            let res = toml::from_str::<v7::Config>(&content)?;
+            let wrong_keys = v7::wrong_keys(&table);
             for key in wrong_keys.iter() {
                 tracing::warn!(%key, "ignoring unsupported key in configuration file");
             }
