+++ gnosis_vpn-lib/src/connection/down/runner.rs
@@ -100,7 +100,7 @@ async fn open_bridge_session(
     };
     hopr.open_session(
         down.destination.address,
-        options.sessions.bridge.target.clone(),
+        down.destination.bridge_target(&options.sessions.bridge.target),
         Some(1),
         Some(1),
         cfg.clone(),
