+++ gnosis_vpn-lib/src/route_health.rs
@@ -912,7 +912,7 @@ impl HealthSession {
         let meta = hopr
             .open_session(
                 destination.address,
-                options.sessions.bridge.target.clone(),
+                destination.bridge_target(&options.sessions.bridge.target),
                 None,
                 None,
                 cfg,
@@ -1193,13 +1193,16 @@ mod tests {
     // --- failure_backoff ---
 
     fn backoff_at(failures: u32) -> Duration {
-        use crate::connection::destination::{Destination, HopRouting};
+        use crate::connection::destination::{Destination, DestinationSource, HopRouting};
         use tokio_util::sync::CancellationToken;
         let dest = Destination::new(
             "test".to_string(),
             addr(1),
             HopRouting::try_from(1).unwrap(),
             Default::default(),
+            None,
+            None,
+            DestinationSource::Configured,
         );
         let mut rh = RouteHealth::new(&dest, false, false, CancellationToken::new());
         rh.exit_failures = failures;
